+++ src/ott/neural/methods/neuraldual.py
@@ -11,10 +11,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import abc
 import warnings
 from typing import (
-    Any,
     Callable,
     Dict,
     Iterator,
@@ -28,138 +26,18 @@ from typing import (
 import jax
 import jax.numpy as jnp
 
-import flax.linen as nn
 import optax
-from flax import struct
-from flax.core import frozen_dict
-from flax.training import train_state
 
 from ott import utils
 from ott.geometry import costs
-from ott.neural.duality import conjugate, models
-from ott.problems.linear import potentials
+from ott.neural.networks import icnn, potentials
+from ott.neural.networks.layers import conjugate
+from ott.problems.linear import potentials as dual_potentials
 
-__all__ = ["W2NeuralTrainState", "BaseW2NeuralDual", "W2NeuralDual"]
+__all__ = ["W2NeuralDual"]
 
 Train_t = Dict[Literal["train_logs", "valid_logs"], Dict[str, List[float]]]
-Callback_t = Callable[[int, potentials.DualPotentials], None]
-
-PotentialValueFn_t = Callable[[jnp.ndarray], jnp.ndarray]
-PotentialGradientFn_t = Callable[[jnp.ndarray], jnp.ndarray]
-
-
-class W2NeuralTrainState(train_state.TrainState):
-  """Adds information about the model's value and gradient to the state.
-
-  This extends :class:`~flax.training.train_state.TrainState` to include
-  the potential methods from the
-  :class:`~ott.neural.duality.neuraldual.BaseW2NeuralDual` used during training.
-
-  Args:
-    potential_value_fn: the potential's value function
-    potential_gradient_fn: the potential's gradient function
-  """
-  potential_value_fn: Callable[
-      [frozen_dict.FrozenDict[str, jnp.ndarray], Optional[PotentialValueFn_t]],
-      PotentialValueFn_t] = struct.field(pytree_node=False)
-  potential_gradient_fn: Callable[[frozen_dict.FrozenDict[str, jnp.ndarray]],
-                                  PotentialGradientFn_t] = struct.field(
-                                      pytree_node=False
-                                  )
-
-
-class BaseW2NeuralDual(abc.ABC, nn.Module):
-  """Base class for the neural solver models."""
-
-  @property
-  @abc.abstractmethod
-  def is_potential(self) -> bool:
-    """Indicates if the module implements a potential value or a vector field.
-
-    Returns:
-      ``True`` if the module defines a potential, ``False`` if it defines a
-       vector field.
-    """
-
-  def potential_value_fn(
-      self,
-      params: frozen_dict.FrozenDict[str, jnp.ndarray],
-      other_potential_value_fn: Optional[PotentialValueFn_t] = None,
-  ) -> PotentialValueFn_t:
-    r"""Return a function giving the value of the potential.
-
-    Applies the module if :attr:`is_potential` is ``True``, otherwise
-    constructs the value of the potential from the gradient with
-
-    .. math::
-
-      g(y) = -f(\nabla_y g(y)) + y^T \nabla_y g(y)
-
-    where :math:`\nabla_y g(y)` is detached for the envelope theorem
-    :cite:`danskin:67,bertsekas:71`
-    to give the appropriate first derivatives of this construction.
-
-    Args:
-      params: parameters of the module
-      other_potential_value_fn: function giving the value of the other
-        potential. Only needed when :attr:`is_potential` is ``False``.
-
-    Returns:
-      A function that can be evaluated to obtain a potential value, or a linear
-      interpolation of a potential.
-    """
-    if self.is_potential:
-      return lambda x: self.apply({"params": params}, x)
-
-    assert other_potential_value_fn is not None, \
-      "The value of the gradient-based potential depends " \
-      "on the value of the other potential."
-
-    def value_fn(x: jnp.ndarray) -> jnp.ndarray:
-      squeeze = x.ndim == 1
-      if squeeze:
-        x = jnp.expand_dims(x, 0)
-      grad_g_x = jax.lax.stop_gradient(self.apply({"params": params}, x))
-      value = -other_potential_value_fn(grad_g_x) + \
-              jax.vmap(jnp.dot)(grad_g_x, x)
-      return value.squeeze(0) if squeeze else value
-
-    return value_fn
-
-  def potential_gradient_fn(
-      self,
-      params: frozen_dict.FrozenDict[str, jnp.ndarray],
-  ) -> PotentialGradientFn_t:
-    """Return a function returning a vector or the gradient of the potential.
-
-    Args:
-      params: parameters of the module
-
-    Returns:
-      A function that can be evaluated to obtain the potential's gradient
-    """
-    if self.is_potential:
-      return jax.vmap(jax.grad(self.potential_value_fn(params)))
-    return lambda x: self.apply({"params": params}, x)
-
-  def create_train_state(
-      self,
-      rng: jax.Array,
-      optimizer: optax.OptState,
-      input: Union[int, Tuple[int, ...]],
-      **kwargs: Any,
-  ) -> W2NeuralTrainState:
-    """Create initial training state."""
-    params = self.init(rng, jnp.ones(input))["params"]
-
-    return W2NeuralTrainState.create(
-        apply_fn=self.apply,
-        params=params,
-        tx=optimizer,
-        potential_value_fn=self.potential_value_fn,
-        potential_gradient_fn=self.potential_gradient_fn,
-        **kwargs,
-    )
+Callback_t = Callable[[int, dual_potentials.DualPotentials], None]
 
 
 class W2NeuralDual:
@@ -228,8 +106,8 @@ class W2NeuralDual:
   def __init__(
       self,
       dim_data: int,
-      neural_f: Optional[BaseW2NeuralDual] = None,
-      neural_g: Optional[BaseW2NeuralDual] = None,
+      neural_f: Optional[potentials.BasePotential] = None,
+      neural_g: Optional[potentials.BasePotential] = None,
       optimizer_f: Optional[optax.OptState] = None,
       optimizer_g: Optional[optax.OptState] = None,
       num_train_iters: int = 20000,
@@ -266,9 +144,9 @@ class W2NeuralDual:
 
     # set default neural architectures
     if neural_f is None:
-      neural_f = models.ICNN(dim_data=dim_data, dim_hidden=[64, 64, 64, 64])
+      neural_f = icnn.ICNN(dim_data=dim_data, dim_hidden=[64, 64, 64, 64])
     if neural_g is None:
-      neural_g = models.ICNN(dim_data=dim_data, dim_hidden=[64, 64, 64, 64])
+      neural_g = icnn.ICNN(dim_data=dim_data, dim_hidden=[64, 64, 64, 64])
     self.neural_f = neural_f
     self.neural_g = neural_g
 
@@ -285,8 +163,8 @@ class W2NeuralDual:
   def setup(
       self,
       rng: jax.Array,
-      neural_f: BaseW2NeuralDual,
-      neural_g: BaseW2NeuralDual,
+      neural_f: potentials.BasePotential,
+      neural_g: potentials.BasePotential,
       dim_data: int,
       optimizer_f: optax.OptState,
       optimizer_g: optax.OptState,
@@ -301,13 +179,13 @@ class W2NeuralDual:
         f"the `W2NeuralDual` setting, with positive weights " \
         f"being {self.pos_weights}."
     if isinstance(
-        neural_f, models.ICNN
+        neural_f, icnn.ICNN
     ) and neural_f.pos_weights is not self.pos_weights:
       warnings.warn(warn_str, stacklevel=2)
       neural_f.pos_weights = self.pos_weights
 
     if isinstance(
-        neural_g, models.ICNN
+        neural_g, icnn.ICNN
     ) and neural_g.pos_weights is not self.pos_weights:
       warnings.warn(warn_str, stacklevel=2)
       neural_g.pos_weights = self.pos_weights
@@ -325,7 +203,7 @@ class W2NeuralDual:
 
     # default to using back_and_forth with the non-convex models
     if self.back_and_forth is None:
-      self.back_and_forth = isinstance(neural_f, models.PotentialMLP)
+      self.back_and_forth = isinstance(neural_f, potentials.PotentialMLP)
 
     if self.num_inner_iters == 1 and self.parallel_updates:
       self.train_step_parallel = self.get_step_fn(
@@ -359,8 +237,8 @@ class W2NeuralDual:
       validloader_source: Iterator[jnp.ndarray],
       validloader_target: Iterator[jnp.ndarray],
       callback: Optional[Callback_t] = None,
-  ) -> Union[potentials.DualPotentials, Tuple[potentials.DualPotentials,
-                                              Train_t]]:
+  ) -> Union[dual_potentials.DualPotentials,
+             Tuple[dual_potentials.DualPotentials, Train_t]]:
     logs = self.train_fn(
         trainloader_source,
         trainloader_target,
@@ -643,7 +521,7 @@ class W2NeuralDual:
 
   def to_dual_potentials(
       self, finetune_g: bool = True
-  ) -> potentials.DualPotentials:
+  ) -> dual_potentials.DualPotentials:
     r"""Return the Kantorovich dual potentials from the trained potentials.
 
     Args:
@@ -664,7 +542,7 @@ class W2NeuralDual:
       )
       return -f_value(grad_g_y) + jnp.dot(grad_g_y, y)
 
-    return potentials.DualPotentials(
+    return dual_potentials.DualPotentials(
         f=f_value,
         g=g_value_prediction if not finetune_g or self.conjugate_solver is None
         else g_value_finetuned,
