+++ src/ott/neural/networks/layers/__init__.py
@@ -11,4 +11,4 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from . import map_estimator, monge_gap
+from . import conjugate, posdef, time_encoder
