+++ src/ott/solvers/__init__.py
@@ -11,4 +11,4 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from . import linear, quadratic, was_solver
+from . import linear, quadratic, utils, was_solver
