+++ src/ott/solvers/utils.py
@@ -24,7 +24,6 @@ __all__ = [
     "match_quadratic",
     "sample_joint",
     "sample_conditional",
-    "cyclical_time_encoder",
     "uniform_sampler",
 ]
 
@@ -148,24 +147,6 @@ def sample_conditional(
 
   src_ixs = jnp.repeat(indices[:, None], k, axis=1)  # (n, k)
   return src_ixs, tgt_ixs
-
-
-def cyclical_time_encoder(t: jnp.ndarray, n_freqs: int = 128) -> jnp.ndarray:
-  r"""Encode time :math:`t` into a cyclical representation.
-
-  Time :math:`t` is encoded as :math:`cos(\hat{t})` and :math:`sin(\hat{t})`
-  where :math:`\hat{t} = [2\pi t, 2\pi 2 t,\dots, 2\pi n_f t]`.
-
-  Args:
-    t: Time of shape ``[n, 1]``.
-    n_freqs: Frequency :math:`n_f` of the cyclical encoding.
-
-  Returns:
-    Encoded time of shape ``[n, 2 * n_freqs]``.
-  """
-  freq = 2 * jnp.arange(n_freqs) * jnp.pi
-  t = freq * t
-  return jnp.concatenate([jnp.cos(t), jnp.sin(t)], axis=-1)
 
 
 def uniform_sampler(
