+++ tests/geometry/graph_test.py
@@ -20,9 +20,9 @@ from networkx.generators import balanced_tree, random_graphs
 import pytest
 
 import jax
+import jax.experimental.sparse as jesp
 import jax.numpy as jnp
 import numpy as np
-from jax.experimental import sparse
 
 from ott.geometry import geometry, graph
 from ott.problems.linear import linear_problem
@@ -259,7 +259,7 @@ class TestGraph:
         data: jnp.ndarray, rows: jnp.ndarray, cols: jnp.ndarray,
         shape: Tuple[int, int]
     ) -> float:
-      G = sparse.BCOO((data, jnp.c_[rows, cols]), shape=shape).todense()
+      G = jesp.BCOO((data, jnp.c_[rows, cols]), shape=shape).todense()
 
       geom = graph.Graph.from_graph(G, t=1.0)
       solver = sinkhorn.Sinkhorn(lse_mode=False, **kwargs)
@@ -274,7 +274,7 @@ class TestGraph:
 
     eps = 1e-3
     G = random_graph(20, p=0.5)
-    G = sparse.BCOO.fromdense(G)
+    G = jesp.BCOO.fromdense(G)
 
     w, rows, cols = G.data, G.indices[:, 0], G.indices[:, 1]
     v_w = jax.random.normal(rng, shape=w.shape)
