+++ tests/initializers/neural/meta_initializer_test.py
@@ -18,11 +18,11 @@ import pytest
 import jax
 import jax.numpy as jnp
 
-import flax.linen as nn
+from flax import linen as nn
 
 from ott.geometry import pointcloud
 from ott.initializers.linear import initializers as linear_init
-from ott.neural.duality import models as nn_init
+from ott.initializers.neural import meta_initializer as meta_init
 from ott.problems.linear import linear_problem
 from ott.solvers.linear import sinkhorn
 
@@ -109,7 +109,7 @@ class TestMetaInitializer:
 
     # overfit the initializer to the problem.
     meta_model = MetaMLP(n)
-    meta_initializer = nn_init.MetaInitializer(geom, meta_model)
+    meta_initializer = meta_init.MetaInitializer(geom, meta_model)
     for _ in range(50):
       _, _, meta_initializer.state = meta_initializer.update(
           meta_initializer.state, a=a, b=b
