+++ tests/neural/methods/monge_gap_test.py
@@ -11,14 +11,19 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from typing import Optional
+
 import pytest
 
 import jax
+import jax.numpy as jnp
 import numpy as np
 
-from ott.geometry import costs
-from ott.neural.duality import models
-from ott.neural.gaps import monge_gap
+from ott import datasets
+from ott.geometry import costs, pointcloud
+from ott.neural.methods import monge_gap
+from ott.neural.networks import potentials
+from ott.tools import sinkhorn_divergence
 
 
 @pytest.mark.fast()
@@ -34,7 +39,7 @@ class TestMongeGap:
     rng1, rng2 = jax.random.split(rng, 2)
     reference_points = jax.random.normal(rng1, (n_samples, n_features))
 
-    model = models.PotentialMLP(dim_hidden=[8, 8], is_potential=False)
+    model = potentials.PotentialMLP(dim_hidden=[8, 8], is_potential=False)
     params = model.init(rng2, x=reference_points[0])
     target = model.apply(params, reference_points)
 
@@ -122,3 +127,67 @@ class TestMongeGap:
     np.testing.assert_array_equal(
         np.isfinite(monge_gap_from_samples_value_cost_fn), True
     )
+
+
+@pytest.mark.fast()
+class TestMongeGapEstimator:
+
+  def test_map_estimator_convergence(self):
+    """Tests convergence of a simple
+    map estimator with Sinkhorn divergence fitting loss
+    and Monge (coupling) gap regularizer.
+    """
+
+    # define the fitting loss and the regularizer
+    def fitting_loss(
+        samples: jnp.ndarray,
+        mapped_samples: jnp.ndarray,
+    ) -> Optional[float]:
+      r"""Sinkhorn divergence fitting loss."""
+      div = sinkhorn_divergence.sinkhorn_divergence(
+          pointcloud.PointCloud,
+          x=samples,
+          y=mapped_samples,
+      ).divergence
+      return div, None
+
+    def regularizer(x, y):
+      gap, out = monge_gap.monge_gap_from_samples(x, y, return_output=True)
+      return gap, out.n_iters
+
+    # define the model
+    model = potentials.PotentialMLP(dim_hidden=[16, 8], is_potential=False)
+
+    # generate data
+    train_dataset, valid_dataset, dim_data = (
+        datasets.create_gaussian_mixture_samplers(
+            name_source="simple",
+            name_target="circle",
+            train_batch_size=30,
+            valid_batch_size=30,
+        )
+    )
+
+    # fit the map
+    solver = monge_gap.MongeGapEstimator(
+        dim_data=dim_data,
+        fitting_loss=fitting_loss,
+        regularizer=regularizer,
+        model=model,
+        regularizer_strength=1.0,
+        num_train_iters=15,
+        logging=True,
+        valid_freq=5,
+    )
+    neural_state, logs = solver.train_map_estimator(
+        *train_dataset, *valid_dataset
+    )
+
+    # check if the loss has decreased during training
+    assert logs["train"]["total_loss"][0] > logs["train"]["total_loss"][-1]
+
+    # check dimensionality of the mapped source
+    source = next(train_dataset.source_iter)
+    mapped_source = neural_state.apply_fn({"params": neural_state.params},
+                                          source)
+    assert mapped_source.shape[1] == dim_data
